RUMM-2401 Move datadog provider in own folder
