+++ packages/core/release-content.txt
@@ -74,8 +74,9 @@ package/ios/Sources/QueueHolder.swift
 package/ios/Sources/RNDdSdkConfiguration.swift
 package/lib
 package/lib/commonjs
-package/lib/commonjs/DatadogProvider.js
-package/lib/commonjs/DatadogProvider.js.map
+package/lib/commonjs/DatadogProvider
+package/lib/commonjs/DatadogProvider/DatadogProvider.js
+package/lib/commonjs/DatadogProvider/DatadogProvider.js.map
 package/lib/commonjs/DdSdkReactNative.js
 package/lib/commonjs/DdSdkReactNative.js.map
 package/lib/commonjs/DdSdkReactNativeConfiguration.js
@@ -144,8 +145,9 @@ package/lib/commonjs/types.js.map
 package/lib/commonjs/version.js
 package/lib/commonjs/version.js.map
 package/lib/module
-package/lib/module/DatadogProvider.js
-package/lib/module/DatadogProvider.js.map
+package/lib/module/DatadogProvider
+package/lib/module/DatadogProvider/DatadogProvider.js
+package/lib/module/DatadogProvider/DatadogProvider.js.map
 package/lib/module/DdSdkReactNative.js
 package/lib/module/DdSdkReactNative.js.map
 package/lib/module/DdSdkReactNativeConfiguration.js
@@ -214,7 +216,8 @@ package/lib/module/types.js.map
 package/lib/module/version.js
 package/lib/module/version.js.map
 package/lib/typescript
-package/lib/typescript/DatadogProvider.d.ts
+package/lib/typescript/DatadogProvider
+package/lib/typescript/DatadogProvider/DatadogProvider.d.ts
 package/lib/typescript/DdSdkReactNative.d.ts
 package/lib/typescript/DdSdkReactNativeConfiguration.d.ts
 package/lib/typescript/InternalLog.d.ts
@@ -254,7 +257,12 @@ package/lib/typescript/types.d.ts
 package/lib/typescript/version.d.ts
 package/package.json
 package/src
-package/src/DatadogProvider.tsx
+package/src/DatadogProvider
+package/src/DatadogProvider/DatadogProvider.tsx
+package/src/DatadogProvider/__tests__
+package/src/DatadogProvider/__tests__/__utils__
+package/src/DatadogProvider/__tests__/__utils__/renderWithProvider.tsx
+package/src/DatadogProvider/__tests__/initialization.test.tsx
 package/src/DdSdkReactNative.tsx
 package/src/DdSdkReactNativeConfiguration.tsx
 package/src/InternalLog.tsx
@@ -266,10 +274,6 @@ package/src/TrackingConsent.tsx
 package/src/__mocks__
 package/src/__mocks__/InternalLog.tsx
 package/src/__tests__
-package/src/__tests__/DatadogProvider
-package/src/__tests__/DatadogProvider/__utils__
-package/src/__tests__/DatadogProvider/__utils__/renderWithProvider.tsx
-package/src/__tests__/DatadogProvider/initialization.test.tsx
 package/src/__tests__/DdRumWrapper.test.tsx
 package/src/__tests__/DdSdkReactNative.test.tsx
 package/src/__tests__/InternalLog.test.tsx
