+++ packages/core/src/DatadogProvider/__tests__/initialization.test.tsx
